Refactor: utils module
