+++ src/color.rs
@@ -5,9 +5,8 @@ use ansi_term::Color;
 use lazy_static::lazy_static;
 use syntect::highlighting::Color as SyntectColor;
 
-use crate::bat_utils::terminal::to_ansi_color;
 use crate::fatal;
-use crate::syntect_utils;
+use crate::utils;
 
 pub fn parse_color(s: &str, true_color: bool) -> Option<Color> {
     if s == "normal" {
@@ -21,11 +20,11 @@ pub fn parse_color(s: &str, true_color: bool) -> Option<Color> {
     } else {
         s.parse::<u8>()
             .ok()
-            .and_then(syntect_utils::syntect_color_from_ansi_number)
-            .or_else(|| syntect_utils::syntect_color_from_ansi_name(s))
+            .and_then(utils::syntect::syntect_color_from_ansi_number)
+            .or_else(|| utils::syntect::syntect_color_from_ansi_name(s))
             .unwrap_or_else(die)
     };
-    to_ansi_color(syntect_color, true_color)
+    utils::bat::terminal::to_ansi_color(syntect_color, true_color)
 }
 
 pub fn color_to_string(color: Color) -> String {
