+++ src/handlers/blame.rs
@@ -73,7 +73,7 @@ impl<'a> StateMachine<'a> {
 
                 // Emit syntax-highlighted code
                 if matches!(self.state, State::Unknown) {
-                    if let Some(lang) = utils::git_blame_filename_extension()
+                    if let Some(lang) = utils::process::git_blame_filename_extension()
                         .as_ref()
                         .or_else(|| self.config.default_language.as_ref())
                     {
