+++ src/utils/bat/terminal.rs
