+++ src/utils/mod.rs
@@ -1,0 +1,4 @@
+#[cfg(not(tarpaulin_include))]
+pub mod bat;
+pub mod process;
+pub mod syntect;
